Fix transaction hooks and improve transactions loading experience in mobile
