+++ packages/desktop-client/src/hooks/useTransactions.ts
@@ -1,12 +1,23 @@
-import { useState, useRef, useEffect, useCallback } from 'react';
+import { useEffect, useRef, useState, useCallback } from 'react';
 
-import type { Query } from 'loot-core/shared/query';
-import type { TransactionEntity } from 'loot-core/types/models';
+import { type Query } from 'loot-core/shared/query';
+import { type IntegerAmount } from 'loot-core/shared/util';
+import { type TransactionEntity } from 'loot-core/types/models';
 
 import {
-  type PagedQuery,
   pagedQuery,
+  type PagedQuery,
 } from '@desktop-client/queries/pagedQuery';
+
+// Mirrors the `splits` AQL option from the server
+type TransactionSplitsOption = 'all' | 'inline' | 'grouped' | 'none';
+
+type CalculateRunningBalancesOption =
+  | ((
+      transactions: TransactionEntity[],
+      splits: TransactionSplitsOption,
+    ) => Map<TransactionEntity['id'], IntegerAmount>)
+  | boolean;
 
 type UseTransactionsProps = {
   /**
@@ -14,22 +25,72 @@ type UseTransactionsProps = {
    * to prevent unnecessary re-renders i.e. `useMemo`, `useState`, etc.
    */
   query?: Query;
+  /**
+   * The options to configure the hook behavior.
+   */
   options?: {
+    /**
+     * The number of transactions to load at a time.
+     * This is used for pagination and should be set to a reasonable number
+     * to avoid loading too many transactions at once.
+     * The default is 50.
+     * @default 50
+     */
     pageCount?: number;
+    /**
+     * Whether to calculate running balances for the transactions returned by the query.
+     * This can be set to `true` to calculate running balances for all transactions
+     * (using the default running balance calculation), or a function that takes the
+     * transactions and the query state and returns a map of transaction IDs to running balances.
+     * The function will be called with the transactions and the query state
+     * whenever the transactions are loaded or reloaded.
+     *
+     * The default running balance calculation is a simple sum of the transaction amounts
+     * in reverse order (bottom up). This works well if the transactions are ordered by
+     * date in descending order. If the query orders the transactions differently,
+     * a custom `calculateRunningBalances` function should be used instead.
+     * @default false
+     */
+    calculateRunningBalances?: CalculateRunningBalancesOption;
   };
 };
+
 type UseTransactionsResult = {
+  /**
+   * The transactions returned by the query.
+   */
   transactions: ReadonlyArray<TransactionEntity>;
+  /**
+   * The running balances for the transactions returned by the query.
+   * This is only populated if `calculateRunningBalances` is either set to `true`
+   * or a function that implements the calculation in the options.
+   */
+  runningBalances: Map<TransactionEntity['id'], IntegerAmount>;
+  /**
+   * Whether the transactions are currently being loaded.
+   */
   isLoading: boolean;
+  /**
+   * An error that occurred while loading the transactions.
+   */
   error?: Error;
+  /**
+   * Reload the transactions.
+   */
   reload: () => void;
+  /**
+   * Load more transactions.
+   */
   loadMore: () => void;
+  /**
+   * Whether more transactions are currently being loaded.
+   */
   isLoadingMore: boolean;
 };
 
 export function useTransactions({
   query,
-  options = { pageCount: 50 },
+  options = { pageCount: 50, calculateRunningBalances: false },
 }: UseTransactionsProps): UseTransactionsResult {
   const [isLoading, setIsLoading] = useState(true);
   const [isLoadingMore, setIsLoadingMore] = useState(false);
@@ -37,6 +98,9 @@ export function useTransactions({
   const [transactions, setTransactions] = useState<
     ReadonlyArray<TransactionEntity>
   >([]);
+  const [runningBalances, setRunningBalances] = useState<
+    Map<TransactionEntity['id'], IntegerAmount>
+  >(new Map());
 
   const pagedQueryRef = useRef<PagedQuery<TransactionEntity> | null>(null);
 
@@ -73,6 +137,19 @@ export function useTransactions({
       onData: data => {
         if (!isUnmounted) {
           setTransactions(data);
+
+          const calculateFn = getCalculateRunningBalancesFn(
+            optionsRef.current?.calculateRunningBalances,
+          );
+          if (calculateFn) {
+            setRunningBalances(
+              calculateFn(
+                data,
+                query.state.tableOptions?.splits as TransactionSplitsOption,
+              ),
+            );
+          }
+
           setIsLoading(false);
         }
       },
@@ -109,10 +186,59 @@ export function useTransactions({
 
   return {
     transactions,
+    runningBalances,
     isLoading,
     ...(error && { error }),
     reload,
     loadMore,
     isLoadingMore,
   };
+}
+
+function getCalculateRunningBalancesFn(
+  calculateRunningBalances: CalculateRunningBalancesOption = false,
+) {
+  return calculateRunningBalances === true
+    ? calculateRunningBalancesBottomUp
+    : typeof calculateRunningBalances === 'function'
+      ? calculateRunningBalances
+      : undefined;
+}
+
+export function calculateRunningBalancesBottomUp(
+  transactions: TransactionEntity[],
+  splits: TransactionSplitsOption,
+  startingBalance: IntegerAmount = 0,
+) {
+  return (
+    transactions
+      .filter(t => {
+        switch (splits) {
+          case 'all':
+            // Only calculate parent/non-split amounts
+            return !t.parent_id;
+          default:
+            // inline
+            // grouped
+            // none
+            return true;
+        }
+      })
+      // We're using `reduceRight` here to calculate the running balance in reverse order (bottom up).
+      .reduceRight((acc, transaction, index, arr) => {
+        const previousTransactionIndex = index + 1;
+        if (previousTransactionIndex >= arr.length) {
+          // This is the last transaction in the list,
+          // so we set the running balance to the starting balance + the amount of the transaction
+          acc.set(transaction.id, startingBalance + transaction.amount);
+          return acc;
+        }
+        const previousTransaction = arr[previousTransactionIndex];
+        const previousRunningBalance = acc.get(previousTransaction.id) ?? 0;
+        const currentRunningBalance =
+          previousRunningBalance + transaction.amount;
+        acc.set(transaction.id, currentRunningBalance);
+        return acc;
+      }, new Map<TransactionEntity['id'], IntegerAmount>())
+  );
 }
